requirements-build.txt
pyinstaller>=6.0
Pillow>=10.0
